+++ ts/features/itwallet/common/store/selectors/remoteConfig.ts
@@ -108,14 +108,15 @@ export const itwDeferredIssuanceScreenContentSelector = createSelector(
 );
 
 /**
- * Return the remote config about iPatente CTA inside the MDL credential details screen.
+ * Return the remote config content for the deferred issuance screen content.
  */
-export const itwIPatenteCtaConfigSelector = createSelector(
+export const itwIsIPatenteCtaEnabledSelector = createSelector(
   itwRemoteConfigSelector,
   itwConfig =>
     pipe(
       itwConfig,
-      O.map(itw => itw.ipatente_cta),
-      O.toUndefined
+      O.map(itw => itw.ipatente_cta_visible),
+      O.getOrElse(() => false)
     )
 );
+