+++ ts/features/pushNotifications/analytics/index.ts
@@ -90,3 +90,69 @@ export const trackNotificationPermissionsStatus = (
   });
   void mixpanelTrack(eventName, props);
 };
+
+export const trackPushNotificationsBannerVisualized = (route: string) => {
+  const eventName = "BANNER";
+  const props = buildEventProperties("UX", "screen_view", {
+    banner_id: "push_notif_activation",
+    banner_page: route,
+    banner_landing: "os_notification_settings"
+  });
+  void mixpanelTrack(eventName, props);
+};
+
+export const trackPushNotificationsBannerTap = (route: string) => {
+  const eventName = "TAP_BANNER";
+  const props = buildEventProperties("UX", "action", {
+    banner_id: "push_notif_activation",
+    banner_page: route,
+    banner_landing: "os_notification_settings"
+  });
+  void mixpanelTrack(eventName, props);
+};
+
+export const trackPushNotificationsBannerClosure = (route: string) => {
+  const eventName = "CLOSE_BANNER";
+  const props = buildEventProperties("UX", "action", {
+    banner_id: "push_notif_activation",
+    banner_page: route,
+    banner_landing: "os_notification_settings"
+  });
+  void mixpanelTrack(eventName, props);
+};
+
+export const trackPushNotificationSystemPopupShown = () => {
+  const eventName = "PUSH_NOTIF_SYSTEM_ALERT";
+  const props = buildEventProperties("UX", "screen_view");
+  void mixpanelTrack(eventName, props);
+};
+
+export const trackPushNotificationBannerDismissAlert = () => {
+  const eventName = "PUSH_NOTIF_THIRD_DISMISS_ALERT";
+  const props = buildEventProperties("UX", "screen_view");
+  void mixpanelTrack(eventName, props);
+};
+
+export const trackPushNotificationBannerDismissOutcome = (
+  outcome: "deactivate" | "remind_later"
+) => {
+  const eventName = "PUSH_NOTIF_THIRD_DISMISS_ALERT_INTERACTION";
+  const props = buildEventProperties("UX", "action", { outcome });
+  void mixpanelTrack(eventName, props);
+};
+
+export const trackPushNotificationBannerForceShow = () => {
+  const eventName = "PUSH_NOTIF_BANNER_FORCE_SHOW";
+  const props = buildEventProperties("TECH", undefined);
+  void mixpanelTrack(eventName, props);
+};
+
+export const trackPushNotificationBannerStillHidden = (
+  unreadMessagesCount: number
+) => {
+  const eventName = "PUSH_NOTIF_BANNER_STILL_HIDDEN";
+  const props = buildEventProperties("TECH", undefined, {
+    unread_count: unreadMessagesCount
+  });
+  void mixpanelTrack(eventName, props);
+};